Add IdleFederationMembersTracker class
