+++ src/Stratis.Bitcoin.Features.PoA/Voting/VotingManager.cs
@@ -49,6 +49,8 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
         /// <summary>All access should be protected by <see cref="locker"/>.</remarks>
         public PollsRepository PollsRepository { get; private set; }
 
+        private IdleFederationMembersTracker idleFederationMembersTracker;
+
         private IIdleFederationMembersKicker idleFederationMembersKicker;
 
         private INodeLifetime nodeLifetime;
@@ -87,6 +89,7 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
             this.votingDataEncoder = new VotingDataEncoder(loggerFactory);
             this.scheduledVotingData = new List<VotingData>();
             this.PollsRepository = new PollsRepository(network, dataFolder, loggerFactory, dBreezeSerializer, chainIndexer);
+            this.idleFederationMembersTracker = new IdleFederationMembersTracker(network, this.PollsRepository, dBreezeSerializer);
             this.logger = loggerFactory.CreateLogger(this.GetType().FullName);
             this.network = network;
             this.poaConsensusOptions = (PoAConsensusOptions)this.network.Consensus.Options;
@@ -497,7 +500,7 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
                 fedMemberKey = member.PubKey;
 
                 // TODO: Remove this debug code. It simulates the header signature rule.
-                if (this.PollsRepository.TryGetLastCachedActivity(fedMemberKey, out (uint blockHeight, uint256 blockHash, uint blockTime, PollsRepository.Activity activity) lastActivity))
+                if (this.idleFederationMembersTracker.TryGetLastCachedActivity(fedMemberKey, out (uint blockHeight, uint256 blockHash, uint blockTime, IdleFederationMembersTracker.Activity activity) lastActivity))
                 {
                     uint idleTimeSeconds = chBlock.ChainedHeader.Header.Time - lastActivity.blockTime;
                     uint expectedIdleTimeSeconds = (uint)modifiedFederation.Count * (uint)this.poaConsensusOptions.TargetSpacingSeconds;
@@ -516,7 +519,7 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
                     }
                 }
                 
-                this.PollsRepository.RecordActivity(transaction, fedMemberKey, (uint)chBlock.ChainedHeader.Height, chBlock.ChainedHeader.HashBlock, PollsRepository.Activity.Mined, chBlock.ChainedHeader.Header.Time);
+                this.idleFederationMembersTracker.RecordActivity(transaction, fedMemberKey, (uint)chBlock.ChainedHeader.Height, chBlock.ChainedHeader.HashBlock, IdleFederationMembersTracker.Activity.Mined, chBlock.ChainedHeader.Header.Time);
 
                 lock (this.locker)
                 {
@@ -536,7 +539,7 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
                         if (poll.VotingData.Key == VoteKey.AddFederationMember)
                         {
                             IFederationMember federationMember = ((PoAConsensusFactory)this.network.Consensus.ConsensusFactory).DeserializeFederationMember(poll.VotingData.Data);
-                            this.PollsRepository.RecordActivity(transaction, federationMember.PubKey, (uint)chBlock.ChainedHeader.Height, chBlock.ChainedHeader.HashBlock, PollsRepository.Activity.Joined, chBlock.ChainedHeader.Header.Time);
+                            this.idleFederationMembersTracker.RecordActivity(transaction, federationMember.PubKey, (uint)chBlock.ChainedHeader.Height, chBlock.ChainedHeader.HashBlock, IdleFederationMembersTracker.Activity.Joined, chBlock.ChainedHeader.Header.Time);
                         }
                     }
                 }
@@ -619,7 +622,7 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
                         // Inactive members don't participate in voting.
                         ChainedHeader pollStartHeader = this.chainIndexer.GetHeader(poll.PollStartBlockData.Hash);
                         var voters = new ConcurrentHashSet<string>(modifiedFederation
-                            .Where(m => ((CollateralFederationMember)m).IsMultisigMember || !this.PollsRepository.IsMemberInactive(transaction, m, chBlock.ChainedHeader))
+                            .Where(m => ((CollateralFederationMember)m).IsMultisigMember || !this.idleFederationMembersTracker.IsMemberInactive(transaction, m, chBlock.ChainedHeader))
                             .Select(m => m.PubKey.ToHex()))
                             .ToArray();
